core: split libvlc loading code paths per platform
